refactor: 일정 목록 조회 시 함께할 친구수 조회 방식을 반정규화에서 서브쿼리로 변경

+++ src/main/java/dev/gmelon/plango/domain/schedule/query/dto/ScheduleListQueryDto.java
@@ -24,7 +24,7 @@ public class ScheduleListQueryDto {
 
     private LocalTime endTime;
 
-    private int memberCount;
+    private long memberCount;
 
     private Boolean isOwner;
